[WIP] setup single pipeline for Postgre, cleanup

+++ ktor-client-cio-mongodb/src/io/ktor/experimental/client/mongodb/util/Util.kt
@@ -48,7 +48,7 @@ fun BytePacketBuilder.writePacketWithIntLength(
     writeFully(bytes)
 }
 
-fun ByteReadPacket.readPacket(count: Int = remaining, order: ByteOrder = ByteOrder.BIG_ENDIAN): ByteReadPacket {
+fun ByteReadPacket.readPacket(count: Int = remaining.toInt(), order: ByteOrder = ByteOrder.BIG_ENDIAN): ByteReadPacket {
     return readBytes(count).asReadPacket(order)
 }
